Flense generator. Reorganize code. Fix for -r 1 hangs.
Flense generator, simplifying away most "event data"
and using what's allocated at the previous stage. In
the future, some data and code can be moved from the
program types to the generator, ending up in-between
the layers as part of pre-allocation code, e.g. move
voicegraph.

The basic main idea is very old, not new. Now the
stubbornly kept premature abstraction of the past
is gone in the generator.

Compared to the flat 2020-06 "mgensys" design, reduces
differences in part. But here, in a more multi-layered
design, there's already specialized data types for the
"updates" to be done, use those instead of duplicating
data with the generator's own types.

Now uses mempool for all allocation in generator.

Reorganize code:
 * Move builder.c into builder/.
 * Rename renderer/ to interp/.
 * Move renderer.c (renamed player.c)
   and the audiodev and wavfile code
   into player/.

Add minimum per-channel buffer length
(currently 1 sample) in player.c, fix
hang when -r 1 looped a 0-length run.

Also change ramp run and skip functions to accept
NULL pos for jumping ahead in time to goal value.

Expand IDEAS file (design ideas) and rename to IDEAS.md.

Change log:
 * 2020-10-24: Silence implicit conversion warning in osc.c,
       add (float) before INT32_MAX.
 * 2020-10-26: Add runalloc module (placeholder)
       and use in generator, as in 'old-dev_202006' branch.
 * 2020-10-27: Make ramp SSG_Ramp_skip()
       skip all remaining values if pos is NULL. Simple but
       may remain unused without a new use-idea.
       Flense generator, simplifying away most "event data"
       and using what's allocated at the previous stage.
       Remove runalloc module for now, for simpler diff.
       Now uses mempool for all allocation in generator.
 * 2020-11-11: Merge part-new code reorganization.
       Use saugns v0.3.5 IDEAS.md with update for changes.
       Adjust SSG_Ramp_run() for consistency when pos NULL.
   Old branch:
    - 2019-12-12: Move builder.c into builder/.
    - 2019-12-15: Move audiodev and wavfile into renderer/.
          Move old files from renderer/ to interp/.
          Expand IDEAS file, rename to IDEAS.md.
    - 2020-07-22: Rename renderer to player.
    - 2020-11-09: Naming in player/player.c.
 * 2020-12-14: Ensure output buffer holds >= 1 sample, fix
       hang with -r 1 when 0 samples were ran in run loop.

+++ interp/osc.c
@@ -35,7 +35,7 @@ void SSG_Osc_run(SSG_Osc *restrict o,
 	for (size_t i = 0; i < buf_len; ++i) {
 		int32_t s_pm = 0;
 		if (pm_f != NULL) {
-			s_pm = lrintf(pm_f[i] * INT32_MAX);
+			s_pm = lrintf(pm_f[i] * (float) INT32_MAX);
 		}
 		float s = SSG_Osc_get(o, freq[i], s_pm) * amp[i];
 		if (layer > 0) s += buf[i];
@@ -62,7 +62,7 @@ void SSG_Osc_run_env(SSG_Osc *restrict o,
 	for (size_t i = 0; i < buf_len; ++i) {
 		int32_t s_pm = 0;
 		if (pm_f != NULL) {
-			s_pm = lrintf(pm_f[i] * INT32_MAX);
+			s_pm = lrintf(pm_f[i] * (float) INT32_MAX);
 		}
 		float s = SSG_Osc_get(o, freq[i], s_pm);
 		float s_amp = amp[i] * 0.5f;
